+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Revamped the ingame cheat menu, it is now a "command menu" which will house all added user commands. It is now accessible in all games not just ones with cheats mode on.
 - Optimized the ingame hero builder menu, should no longer cause instability or lags. 
 - Added a ingame filter to see hero perk related abiliites. 
+- Added cheat '-wtf' that can be used in games without sv_cheats
 
 ## 6.9 - Steam Version
 - You can enable cheats in a multiplayer games now if everbody types '-enablecheats'